Modulando/organizando o programa principal e desenvolvendo a classe CreateTemplate

+++ Commands/DeleteFile.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using DirectoryCLI.CommandStyles;
+using DirectoryCLI.Exceptions;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -10,16 +12,44 @@ namespace DirectoryCLI.Commands
 {
     internal class DeleteFile : CommandsConfig
     {
-        public string ArchiveName { get; }
+        static  string ArchiveName { get; set; }
 
         public DeleteFile(string archiveName)
         {
             ArchiveName = archiveName;
         }
-        public void Execute(FileInfo directoryPath)
+        public static void Execute(FileInfo directoryPath, string[] arguments)
         {
-            string fullPath = Path.Combine(directoryPath.FullName, ArchiveName);
-            File.Delete(fullPath);
+            for (int i = 0; i <= arguments.Length - 3; i++)
+            {
+                try
+                {
+                    if (File.Exists(Path.Combine(directoryPath.FullName, arguments[2 + i])))
+                    {
+                        string fullPath = Path.Combine(directoryPath.FullName, ArchiveName);
+                        File.Delete(fullPath);
+
+                        colors.Red();
+
+                        Console.WriteLine($"Arquivo [{arguments[2 + i]}] deletado em [{arguments[0]}]");
+                    }
+                    else
+                    {
+                        throw new ItemNotFoundException("Não foi possível criar um arquivo: ");
+                    }
+                }
+                catch (ItemNotFoundException ex)
+                {
+                    colors.DarkRed();
+
+                    Console.Write(ex.Message);
+
+                    colors.Red();
+
+                    Console.WriteLine($"O arquivo '{arguments[2 + i]}' não existe neste diretório.");
+                    Colors.WhiteText();
+                }
+            }
         }
     }
 }
